+++ core/src/main/java/inetsoft/web/wiz/service/CreateVsService.java
@@ -111,19 +111,33 @@ public class CreateVsService {
             throw new IllegalStateException("Worksheet has no primary assembly");
          }
 
-         Viewsheet newVs = new Viewsheet(sourceWs);
-         newVs.syncWizData(vs);
-         VSAssembly assembly = createAssembly(newVs, model.getVisualizationType(), title, config, primaryAssembly.getName());
+         // Incremental mode: when runtimeId was supplied by the caller, add the new assembly
+         // to the existing viewsheet rather than replacing it wholesale.
+         final Viewsheet targetVs;
+
+         if(createdRuntimeId) {
+            // Fresh viewsheet: create a new one bound to the source worksheet.
+            targetVs = new Viewsheet(sourceWs);
+            targetVs.syncWizData(vs);
+         }
+         else {
+            // Incremental: reuse the existing viewsheet; update the base entry if needed.
+            targetVs = vs;
+            targetVs.setBaseEntry(sourceWs);
+         }
+
+         VSAssembly assembly = createAssembly(targetVs, model.getVisualizationType(), title, config, primaryAssembly.getName());
 
          if(assembly == null) {
             throw new RuntimeException("Unsupported visualization type: " + model.getVisualizationType());
          }
 
-         newVs.addAssembly(assembly);
+         targetVs.addAssembly(assembly);
          assembly.setPrimary(true);
 
+         // Always call setViewsheet so the sandbox picks up the updated viewsheet object.
          Viewsheet previousVs = rvs.getViewsheet();
-         rvs.setViewsheet(newVs);
+         rvs.setViewsheet(targetVs);
 
          try {
             // Execute the assembly view after the viewsheet is set so that dynamic values are
@@ -160,7 +174,15 @@ public class CreateVsService {
             return result;
          }
          catch(Exception e) {
+            // Rollback: restore the previous viewsheet state.
             rvs.setViewsheet(previousVs);
+
+            if(!createdRuntimeId) {
+               // In incremental mode targetVs == previousVs, so also remove the assembly
+               // that was just added to leave the viewsheet in its original state.
+               previousVs.removeAssembly(assembly.getName());
+            }
+
             throw e;
          }
       }
